fix(coordinator): move per-tool drift check into ToolCallCoordinator.registerTools
The previous per-server drift check lived in TrustedProcess.connectMcpServers,
which is only used by integration tests and the direct-tool-call fallback.
Real CLI sessions go through src/sandbox/index.ts (Code Mode + Docker
CodeModeProxy) and never hit it, so renaming a tool in the on-disk
tool-annotations.json produced silent default-denies instead of a warning.

Move the check into ToolCallCoordinator.registerTools — the single
chokepoint both runtime paths funnel through after listTools() — using
the coordinator's existing toolAnnotations field. Remove the now-dead
copy from TrustedProcess.
